✏️ Fix typos and improve file naming: rename 'challange_dashboard_screen.dart' to 'challenge_dashboard_screen.dart' and correct spelling in README
